+++ lib/src/builder/cube_builder.dart
@@ -11,6 +11,7 @@ typedef AsyncCubeWidgetBuilder<C extends Cube> = Widget Function(
 );
 
 typedef InitCallback<C extends Cube> = Function(C cube);
+typedef CubeWidgetDispose<C extends Cube> = bool Function(C cube);
 
 class CubeBuilder<C extends Cube> extends StatefulWidget {
   const CubeBuilder({
@@ -27,7 +28,7 @@ class CubeBuilder<C extends Cube> extends StatefulWidget {
   final AsyncCubeWidgetBuilder<C> builder;
   final OnActionChanged<C, CubeAction> onAction;
   final InitCallback<C> initState;
-  final VoidCallback dispose;
+  final CubeWidgetDispose<C> dispose;
   final C cube;
 
   @override
@@ -39,10 +40,7 @@ class _CubeBuilderState<C extends Cube> extends State<CubeBuilder> with StateMix
 
   @override
   void initState() {
-    cube = widget.cube;
-    if (cube == null) {
-      cube = Cubes.getDependency();
-    }
+    cube = widget.cube ?? Cubes.getDependency();
     cube.data = widget.initData;
     cube.addOnActionListener(_onAction);
     super.initState();
@@ -52,9 +50,8 @@ class _CubeBuilderState<C extends Cube> extends State<CubeBuilder> with StateMix
 
   @override
   void dispose() {
+    if (cubeWidget.dispose?.call(cube) ?? true) cube.dispose();
     cube.removeOnActionListener(_onAction);
-    cube.dispose();
-    cubeWidget.dispose?.call();
     super.dispose();
   }