+++ lib/src/builder/cube_builder_animation.dart
@@ -23,7 +23,7 @@ class CubeBuilderAnimation<C extends Cube> extends StatefulWidget {
   final AsyncCubeWidgetBuilder<C> builder;
   final OnActionChanged<C, CubeAction> onAction;
   final InitStateWithTickerCallback<C> initState;
-  final VoidCallback dispose;
+  final CubeWidgetDispose<C> dispose;
   final C cube;
 
   @override
@@ -50,9 +50,8 @@ class _CubeBuilderAnimationState<C extends Cube> extends State<CubeBuilderAnimat
 
   @override
   void dispose() {
+    if (cubeWidget.dispose?.call(cube) ?? true) cube.dispose();
     cube.removeOnActionListener(_onAction);
-    cube.dispose();
-    cubeWidget.dispose?.call();
     super.dispose();
   }