+++ lib/src/cube.dart
@@ -1,12 +1,15 @@
 import 'package:cubes/cubes.dart';
 import 'package:cubes/src/util/cube_memory_container.dart';
 import 'package:cubes/src/util/debouncer.dart';
+import 'package:flutter/cupertino.dart';
 
 typedef OnActionChanged<A extends Cube, CubeAction> = void Function(A valueA, CubeAction valueB);
 
 abstract class Cube {
   List<OnActionChanged> _onActionListeners;
   Map<dynamic, Debounce> _debounceMap;
+  Map<ObservableValue, VoidCallback> _listenersObservableMap;
+  OnActionChanged _cubeActionListener;
 
   /// initial data if passed through CubeBuilder
   dynamic data;
@@ -18,6 +21,8 @@ abstract class Cube {
 
   /// called when the cube is destroyed
   void dispose() {
+    _disposeListen();
+    removeOnActionListener(_cubeActionListener);
     CubeMemoryContainer.instance.remove(this);
   }
 
@@ -37,8 +42,10 @@ abstract class Cube {
   }
 
   /// Method to send anything to view
+  @protected
   void onAction(CubeAction action) {
-    _onActionListeners?.forEach((element) => element(this, action));
+    if (_onActionListeners?.isEmpty == true) return;
+    _onActionListeners?.last(this, action);
   }
 
   ///
@@ -47,10 +54,11 @@ abstract class Cube {
   /// Example:
   ///
   ///   runDebounce(
-  //      'increment',
-  //      () => print(count.value),
-  //      duration: Duration(seconds: 1),
-  //    );
+  ///      'increment',
+  ///      () => print(count.value),
+  ///      duration: Duration(seconds: 1),
+  ///    );
+  @protected
   void runDebounce(
     dynamic identify,
     Function call, {
@@ -73,12 +81,36 @@ abstract class Cube {
   }
 
   /// Uses to get the Cube ready in memory by type
+  @protected
   T getCubeIsReady<T extends Cube>() {
     return CubeMemoryContainer.instance.get<T>();
   }
 
   /// Uses to get the Cubes ready in memory by type
+  @protected
   Iterable<T> getCubesAreReady<T extends Cube>() {
     return CubeMemoryContainer.instance.getCubes<T>();
+  }
+
+  /// Uses to listen `ObservableValue` inner Cube
+  @protected
+  void listen<T>(ObservableValue<T> observableValue, ValueChanged<T> listener) {
+    if (_listenersObservableMap == null) _listenersObservableMap = Map();
+    _listenersObservableMap[observableValue] = () => listener(observableValue.value);
+    observableValue.addListener(_listenersObservableMap[observableValue]);
+  }
+
+  /// Remove listeners created on `listen`
+  void _disposeListen() {
+    _listenersObservableMap?.forEach((key, value) {
+      key.removeListener(value);
+    });
+  }
+
+  /// Uses to listen `CubeAction` sended to view
+  @protected
+  void listenActions(ValueChanged<CubeAction> listener) {
+    _cubeActionListener = (cube, action) => listener(action);
+    addOnActionListener(_cubeActionListener);
   }
 }